knife_tracker_backend/requirements.txt
fastapi
uvicorn
motor
pymongo
httpx
python-dotenv
